splitting up test suite, improving documentation, uuid validation, routine to retrieve stac items of existing order, upon submitting order: omit search to ensure provided STAC IDs are valid in conjunction with provided items
