+++ capella_console_client/exceptions.py
@@ -1,5 +1,3 @@
-import httpx
-
 from typing import Dict, Any
