+++ capella_console_client/validate.py
@@ -1,0 +1,11 @@
+import uuid
+
+from typing import no_type_check
+
+
+@no_type_check
+def _validate_uuid(uuid_str: str) -> None:
+    try:
+        uuid.UUID(uuid_str)
+    except ValueError as e:
+        raise ValueError(f"{uuid_str} is not a valid uuid: {e}")
